Backend/requirements.txt
fastapi
uvicorn
python-chess
stockfish